+++ src/features/ingestion/usecases.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Callable, Protocol, Self
 
 from src.features.ingestion import scanners as strategy
-from src.features.ingestion.parser import CurriculumParser
+from src.features.ingestion.parsing import CurriculumParser
 from src.lib.panic import panic
 
 if TYPE_CHECKING:
